Cleaner ChrootBuilder
Better comments and code placement

Test for mount point, file, user, group, etc. before making modification

More clean Echo to inform the user what is going on

+++ ChrootBuilder/ChrootBuilder.bash
@@ -19,11 +19,33 @@
 #
 ###########################################################################
 
-# Constant Variables
+#### Constant Variables ####
 chrootTemplateDir="/srv/chrootTemplate"
 chrootDir="/srv/chroot/"
+chrootGroup=chrooted
+chrootShell=/bin/bash
 
-# Create the chroot directory if it didnt exist
+
+
+#### Création du groupe chrooted et ajout de celui-ci dans les configurations de SSH ####
+if [ -z "$(cat /etc/ssh/sshd_config | grep "Match Group ${chrootGroup}")" ]
+	then
+		echo "${chrootGroup} did not exist. [...]"
+		# Create chroot group
+		addgroup $chrootGroup
+		
+		# Add chroot group to sshd_config
+		echo -e '\n' | sudo tee -a /etc/ssh/sshd_config > /dev/null
+		echo -e "Match Group ${chrootGroup}" | sudo tee -a /etc/ssh/sshd_config > /dev/null
+		echo -e "\tChrootDirectory ${chrootDir}/%u" | sudo tee -a /etc/ssh/sshd_config > /dev/null
+		
+		echo "${chrootGroup} added. [OK]"
+		echo "${chrootGroup} added to sshd_config. [OK]"
+fi
+
+
+
+#### Création du répertoire de stockage des chroot ####
 if [ -d $chrootDir ]
 	then
 		echo "${chrootDir} exist. [OK]"
@@ -32,42 +54,60 @@ if [ -d $chrootDir ]
 		mkdir $chrootDir
 fi
 
-port=$(expr $(ls $chrootDir | wc -l) + 1)
+port=$(expr $(ls $chrootDir | wc -l) + 1) # Identifiant de la chroot
 
-#### Création de l'utilisateur ####
+
+
+####### Création de l'utilisateur #######
 # Demander le nom de l'utilisateur
-read -p "Nom d'utilisateur:" username
-read -s -p "Mot de passe:" password
+read -p "Nom d'utilisateur:" username	# Nom de l'utilisateur
+read -s -p "Mot de passe:" password		# Mot de passe de l'utilisateur
 echo -e '\n'
 
-userDir=$chrootDir/$username
-userHome=$userDir/home/$username
+userDir=$chrootDir/$username 			# Répertoire du chroot de l'utilisateur
 
-if [ -z "$(cat /etc/ssh/sshd_config | grep "Match Group chrooted")" ]
+# Création de l'utilisateur, ajout dans le goupe chroot et changement de son shell pour Bash
+if [ -n "$(cat /etc/passwd | grep "${username}:")" ]
 	then
-		# Created chrooted group
-		addgroup chrooted
+		echo "${username} exist. [...]"
+		usermod -a -G $chrootGroup -s $chrootShell $username # Create user with chrooted group and chroot shell
+		echo "${username} added to ${chrootGroup}. [OK]"
+		echo "${username} shell change to ${chrootShell}. [OK]"
+	else
+		echo "${username} did not exist. [...]"
+		useradd -G $chrootGroup -s $chrootShell $username # Add user to chrooted and change shell to chroot shell
+		echo "${username} added. [OK]"
+fi
+
+# Change the password of the user
+echo -e $username:$password | chpasswd
+
+
+
+#### Supprimer les anciennes installation du chroot de l'utilisateur ####
+if [ -d $userDir ]
+	then
+		echo "${userDir} exist. Remove it (Can take a while) [...]"
+
+		# Remove process management from user chroot
+		if mountpoint -q $userDir/proc
+			then
+				echo "${userDir}/proc mounted. [...]"
+				umount $userDir/proc
+		fi
+		echo "${userDir}/proc unmounted. [OK]"
 		
-		# Add users to sshd_config
-		echo -e '\n' | sudo tee -a /etc/ssh/sshd_config > /dev/null
-		echo -e "Match Group chrooted" | sudo tee -a /etc/ssh/sshd_config > /dev/null
-		echo -e "\tChrootDirectory ${chrootDir}/%u" | sudo tee -a /etc/ssh/sshd_config > /dev/null
+		# Remove user chroot
+		rm -r $userDir
+		
+		echo "${userDir} removed. [OK]"
+	else
+		echo "${userDir} did not exist. [OK]"
 fi
 
 
-# Create user
-useradd $username
-usermod -a -G chrooted -s /bin/bash $username
 
-# Password
-echo -e $username:$password | chpasswd
-
-# Créer son chroot
-umount $userDir/etc/resolv.conf
-umount $userDir/proc
-rm -r $userDir
-
-# Configuration de schroot pour l'utilisateur (Dans /etc/schroot/chroot.d/$username.conf)
+#### Configuration de schroot pour l'utilisateur (Dans /etc/schroot/chroot.d/$username.conf) ####
 echo -e "[${username}]" | tee /etc/schroot/chroot.d/$username.conf > /dev/null
 echo -e "description=Ubuntu Focal for ${username}" | tee -a /etc/schroot/chroot.d/$username.conf > /dev/null
 echo -e "directory=${userDir}" | tee -a /etc/schroot/chroot.d/$username.conf > /dev/null
@@ -76,16 +116,24 @@ echo -e "root-groups=${username}" | tee -a /etc/schroot/chroot.d/$username.conf 
 echo -e "type=directory" | tee -a /etc/schroot/chroot.d/$username.conf > /dev/null
 
 
-#### Installer Ubuntu dans le chroot (schroot) ####
-# Remove process management from Template (Need to be mounted later, not copied)
-umount $chrootTemplateDir/proc
 
-# Copier le Template en tant que home
-echo -e "Copy ${chrootTemplateDir} in ${userDir} (Can take a while) ..."
+#### Installer Ubuntu dans le chroot (Copie du Template) ####
+# Remove process management from Template (Need to be mounted later, not copied)
+if mountpoint -q $chrootTemplateDir/proc
+	then
+		echo "${chrootTemplateDir}/proc mounted. [...]"
+		umount $chrootTemplateDir/proc
+		rm -r $chrootTemplateDir/proc
+fi
+echo "${chrootTemplateDir}/proc unmounted and removed. [OK]"
+
+# Copier le Template en tant que chroot pour l'utilisateur
+echo -e "Copy ${chrootTemplateDir} in ${userDir} (Will take a while) [...]"
 cp -a $chrootTemplateDir $userDir
+echo -e "${chrootTemplateDir} copiedvi in ${userDir} [OK]"
 
 #### Configuration de Ubuntu dans le chroot ####
-# Add internet access (Not working?)
+# Add internet access
 cp --parents /run/systemd/resolve/stub-resolv.conf $userDir
 cp --parents /etc/resolv.conf $userDir
 
@@ -93,10 +141,14 @@ cp --parents /etc/resolv.conf $userDir
 mkdir $userDir/proc
 mount -o bind /proc $userDir/proc
 
+
+
 #### Installation de logiciels dans le chroot ####
 #### Lancer inChroot.bash dans le chroot en root #######################
 cp inChroot.bash $userDir
 chroot $userDir /bin/bash inChroot.bash $username $password $port $(id -u $username) $(id -g $username)
+
+
 
 # Restart SSh service (Apply config)
 /etc/init.d/ssh restart
